# Add the interface for streaming the vectors from java to jni layer with initial capacity