--- a/src/main/java/org/opensearch/knn/jni/JNIService.java
+++ b/src/main/java/org/opensearch/knn/jni/JNIService.java
@@ -235,44 +235,30 @@ public class JNIService {
     }
 
     /**
+     * <p>
+     *  The function is deprecated. Use {@link JNICommons#storeVectorData(long, float[][], long)}
+     * </p>
      * Transfer vectors from Java to native
      *
      * @param vectorsPointer pointer to vectors in native memory. Should be 0 to create vector as well
      * @param trainingData data to be transferred
      * @return pointer to native memory location of training data
      */
+    @Deprecated(since = "2.14.0", forRemoval = true)
     public static long transferVectors(long vectorsPointer, float[][] trainingData) {
         return FaissService.transferVectors(vectorsPointer, trainingData);
     }
 
     /**
+     * <p>
+     *  The function is deprecated. Use {@link JNICommons#freeVectorData(long)}
+     * </p>
      * Free vectors from memory
      *
      * @param vectorsPointer to be freed
      */
+    @Deprecated(since = "2.14.0", forRemoval = true)
     public static void freeVectors(long vectorsPointer) {
         FaissService.freeVectors(vectorsPointer);
-    }
-
-    /**
-     * Experimental: Transfer vectors from Java to native layer. This is the version 2 of transfer vector
-     * functionality. The difference between this and the version 1 is, this version puts vectors at the end rather
-     * than in front. Keeping this name as V2 for now, will come up with better name going forward.
-     * <p>
-     * This is not a production ready function for now. Adding this to ensure that we are able to run atleast 1
-     * micro-benchmarks.
-     * </p>
-     * <p>
-     * TODO: Rename the function
-     * <br>
-     * TODO: Make this function native function and use a common cpp file to host these functions.
-     * </p>
-     * @param vectorsPointer pointer to vectors in native memory. Should be 0 to create vector as well
-     * @param data data to be transferred
-     * @return pointer to native memory location for data
-     *
-     */
-    public static long transferVectorsV2(long vectorsPointer, float[][] data) {
-        return FaissService.transferVectorsV2(vectorsPointer, data);
     }
 }