… jni/include/org_opensearch_knn_jni_JNICommons.h