### Description
Add the interface for streaming the vectors from java to jni layer with initial capacity. The PR will be added in a feature branch.

Along with this change I deprecated the methods in FaissService. Will remove those methods going forward. I also removed the transferVectorsV2 function as it is not required now.
 
### JNI test
```
(base) 00:17 ~/workplace/k-NN/jni (stream-vectors)$ ./bin/jni_test 
Running main() from /Users/navneev/workplace/k-NN/jni/googletest-src/googletest/src/gtest_main.cc
[==========] Running 22 tests from 20 test suites.
[----------] Global test environment set-up.
[----------] 1 test from FaissCreateIndexTest
[ RUN      ] FaissCreateIndexTest.BasicAssertions
[       OK ] FaissCreateIndexTest.BasicAssertions (14 ms)
[----------] 1 test from FaissCreateIndexTest (14 ms total)

[----------] 1 test from FaissCreateIndexFromTemplateTest
[ RUN      ] FaissCreateIndexFromTemplateTest.BasicAssertions
[       OK ] FaissCreateIndexFromTemplateTest.BasicAssertions (5 ms)
[----------] 1 test from FaissCreateIndexFromTemplateTest (5 ms total)

[----------] 3 tests from FaissLoadIndexTest
[ RUN      ] FaissLoadIndexTest.BasicAssertions
[       OK ] FaissLoadIndexTest.BasicAssertions (4 ms)
[ RUN      ] FaissLoadIndexTest.HNSWPQDisableSdcTable
WARNING clustering 256 points to 16 centroids: please provide at least 624 training points
[       OK ] FaissLoadIndexTest.HNSWPQDisableSdcTable (372 ms)
[ RUN      ] FaissLoadIndexTest.IVFPQDisablePrecomputeTable
WARNING clustering 256 points to 16 centroids: please provide at least 624 training points
[       OK ] FaissLoadIndexTest.IVFPQDisablePrecomputeTable (359 ms)
[----------] 3 tests from FaissLoadIndexTest (736 ms total)

[----------] 1 test from FaissQueryIndexTest
[ RUN      ] FaissQueryIndexTest.BasicAssertions
[       OK ] FaissQueryIndexTest.BasicAssertions (5 ms)
[----------] 1 test from FaissQueryIndexTest (5 ms total)

[----------] 1 test from FaissQueryIndexWithFilterTest1435
[ RUN      ] FaissQueryIndexWithFilterTest1435.BasicAssertions
[       OK ] FaissQueryIndexWithFilterTest1435.BasicAssertions (10 ms)
[----------] 1 test from FaissQueryIndexWithFilterTest1435 (10 ms total)

[----------] 1 test from FaissQueryIndexWithParentFilterTest
[ RUN      ] FaissQueryIndexWithParentFilterTest.BasicAssertions
[       OK ] FaissQueryIndexWithParentFilterTest.BasicAssertions (5 ms)
[----------] 1 test from FaissQueryIndexWithParentFilterTest (5 ms total)

[----------] 1 test from FaissFreeTest
[ RUN      ] FaissFreeTest.BasicAssertions
[       OK ] FaissFreeTest.BasicAssertions (0 ms)
[----------] 1 test from FaissFreeTest (0 ms total)

[----------] 1 test from FaissInitLibraryTest
[ RUN      ] FaissInitLibraryTest.BasicAssertions
[       OK ] FaissInitLibraryTest.BasicAssertions (0 ms)
[----------] 1 test from FaissInitLibraryTest (0 ms total)

[----------] 1 test from FaissTrainIndexTest
[ RUN      ] FaissTrainIndexTest.BasicAssertions
[       OK ] FaissTrainIndexTest.BasicAssertions (1 ms)
[----------] 1 test from FaissTrainIndexTest (1 ms total)

[----------] 1 test from FaissCreateHnswSQfp16IndexTest
[ RUN      ] FaissCreateHnswSQfp16IndexTest.BasicAssertions
[       OK ] FaissCreateHnswSQfp16IndexTest.BasicAssertions (5 ms)
[----------] 1 test from FaissCreateHnswSQfp16IndexTest (5 ms total)

[----------] 1 test from FaissIsSharedIndexStateRequired
[ RUN      ] FaissIsSharedIndexStateRequired.BasicAssertions
[       OK ] FaissIsSharedIndexStateRequired.BasicAssertions (0 ms)
[----------] 1 test from FaissIsSharedIndexStateRequired (0 ms total)

[----------] 1 test from FaissInitAndSetSharedIndexState
[ RUN      ] FaissInitAndSetSharedIndexState.BasicAssertions
WARNING clustering 256 points to 16 centroids: please provide at least 624 training points
[       OK ] FaissInitAndSetSharedIndexState.BasicAssertions (360 ms)
[----------] 1 test from FaissInitAndSetSharedIndexState (360 ms total)

[----------] 1 test from IDGrouperBitMapTest
[ RUN      ] IDGrouperBitMapTest.BasicAssertions
[       OK ] IDGrouperBitMapTest.BasicAssertions (0 ms)
[----------] 1 test from IDGrouperBitMapTest (0 ms total)

[----------] 1 test from NmslibIndexWrapperSearchTest
[ RUN      ] NmslibIndexWrapperSearchTest.BasicAssertions
[       OK ] NmslibIndexWrapperSearchTest.BasicAssertions (0 ms)
[----------] 1 test from NmslibIndexWrapperSearchTest (0 ms total)

[----------] 1 test from NmslibCreateIndexTest
[ RUN      ] NmslibCreateIndexTest.BasicAssertions
[       OK ] NmslibCreateIndexTest.BasicAssertions (2 ms)
[----------] 1 test from NmslibCreateIndexTest (2 ms total)

[----------] 1 test from NmslibLoadIndexTest
[ RUN      ] NmslibLoadIndexTest.BasicAssertions
[       OK ] NmslibLoadIndexTest.BasicAssertions (1 ms)
[----------] 1 test from NmslibLoadIndexTest (1 ms total)

[----------] 1 test from NmslibQueryIndexTest
[ RUN      ] NmslibQueryIndexTest.BasicAssertions
[       OK ] NmslibQueryIndexTest.BasicAssertions (2 ms)
[----------] 1 test from NmslibQueryIndexTest (2 ms total)

[----------] 1 test from NmslibFreeTest
[ RUN      ] NmslibFreeTest.BasicAssertions
[       OK ] NmslibFreeTest.BasicAssertions (0 ms)
[----------] 1 test from NmslibFreeTest (0 ms total)

[----------] 1 test from NmslibInitLibraryTest
[ RUN      ] NmslibInitLibraryTest.BasicAssertions
[       OK ] NmslibInitLibraryTest.BasicAssertions (0 ms)
[----------] 1 test from NmslibInitLibraryTest (0 ms total)

[----------] 1 test from CommonsTests
[ RUN      ] CommonsTests.BasicAssertions
[       OK ] CommonsTests.BasicAssertions (0 ms)
[----------] 1 test from CommonsTests (0 ms total)

[----------] Global test environment tear-down
[==========] 22 tests from 20 test suites ran. (1154 ms total)
[  PASSED  ] 22 tests.

```
### Old benchmarks

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:////Users/navneev/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip.htm">
<link rel=File-List
href="file:////Users/navneev/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip_filelist.xml">
</head>

<body link="#0563C1" vlink="#954F72">


Benchmark | (dimension) | (vectorsPerTransfer) | Mode | Cnt | Score | Units | Heap Used (in Mb) | Number of trips to JNI Layer
-- | -- | -- | -- | -- | -- | -- | -- | --
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 128 | 100000 | ss | 3 | 1.206 | s/op | 48.82813 | 10
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 128 | 500000 | ss | 3 | 1.576 | s/op | 244.14063 | 2
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 128 | 1000000 | ss | 3 | 1.591 | s/op | 488.28125 | 1
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 256 | 100000 | ss | 3 | 2.926 | s/op | 97.65625 | 10
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 256 | 500000 | ss | 3 | 3.13 | s/op | 488.28125 | 2
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 256 | 1000000 | ss | 3 | 3.668 | s/op | 976.5625 | 1
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 384 | 100000 | ss | 3 | 4.651 | s/op | 146.48438 | 10
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 384 | 500000 | ss | 3 | 7.108 | s/op | 732.42188 | 2
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 384 | 1000000 | ss | 3 | 8.964 | s/op | 1464.84375 | 1
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 512 | 100000 | ss | 3 | 5.36 | s/op | 195.3125 | 10
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 512 | 500000 | ss | 3 | 9.59 | s/op | 976.5625 | 2
TransferVectorsBenchmarks.transferVectorsInitialCapacity | 512 | 1000000 | ss | 3 | 9.959 | s/op | 1953.125 | 1



</body>

</html>




### New Benchmarks

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:////Users/navneev/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip.htm">
<link rel=File-List
href="file:////Users/navneev/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip_filelist.xml">
</head>

<body link="#0563C1" vlink="#954F72">


Benchmark | (dimension) | (vectorsPerTransfer) | Mode | Cnt | Score | Units | Heap Used (in Mb) | Number of trips to JNI Layer
-- | -- | -- | -- | -- | -- | -- | -- | --
TransferVectorsBenchmarks.transferVectors_withCapacity | 128 | 100000 | ss | 3 | 0.873 | s/op | 48.82813 | 10
TransferVectorsBenchmarks.transferVectors_withCapacity | 128 | 500000 | ss | 3 | 0.874 | s/op | 244.14063 | 2
TransferVectorsBenchmarks.transferVectors_withCapacity | 128 | 1000000 | ss | 3 | 0.868 | s/op | 488.28125 | 1
TransferVectorsBenchmarks.transferVectors_withCapacity | 256 | 100000 | ss | 3 | 1.366 | s/op | 97.65625 | 10
TransferVectorsBenchmarks.transferVectors_withCapacity | 256 | 500000 | ss | 3 | 1.379 | s/op | 488.28125 | 2
TransferVectorsBenchmarks.transferVectors_withCapacity | 256 | 1000000 | ss | 3 | 1.261 | s/op | 976.5625 | 1
TransferVectorsBenchmarks.transferVectors_withCapacity | 384 | 100000 | ss | 3 | 1.773 | s/op | 146.48438 | 10
TransferVectorsBenchmarks.transferVectors_withCapacity | 384 | 500000 | ss | 3 | 1.776 | s/op | 732.42188 | 2
TransferVectorsBenchmarks.transferVectors_withCapacity | 384 | 1000000 | ss | 3 | 1.786 | s/op | 1464.84375 | 1
TransferVectorsBenchmarks.transferVectors_withCapacity | 512 | 100000 | ss | 3 | 2.255 | s/op | 195.3125 | 10
TransferVectorsBenchmarks.transferVectors_withCapacity | 512 | 500000 | ss | 3 | 2.289 | s/op | 976.5625 | 2
TransferVectorsBenchmarks.transferVectors_withCapacity | 512 | 1000000 | ss | 3 | 2.283 | s/op | 1953.125 | 1



</body>

</html>


### Issues Resolved
This is first of the many PR that will be raised for this issue: #1506
 
### Check List
- [X] New functionality includes testing.
  - [X] All tests pass
- [X] New functionality has been documented.
  - [X] New functionality has javadoc added
- [X] Commits are signed as per the DCO using --signoff 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/opensearch-project/k-NN/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).