@martin-gaievski 2024-04-02
> One other thing @navneet1v - I think we might need to update https://github.com/opensearch-project/k-NN/blob/main/src/main/plugin-metadata/plugin-security.policy#L2-L4 with new jni lib.

if this change is needed how does it work today? Looks like Navneet done many benchmarks with current code.