@jmazanec15 2024-04-02
> if this change is needed how does it work today? Looks like Navneet done many benchmarks with current code.

@martin-gaievski  @navneet1v I have a suspicion that faiss is being called first which is then loading the common library in the C++ code - but not sure either how it was working.